Juu yritän alkaa työntämään kaiken materiaalini tänne. -Siinä siis jokunen tekemäni input funktio. -Lisää tulossa.

+++ kysyPosKokonaisluku.c
@@ -10,7 +10,7 @@
 
     (vähän modifoitu tähän tehtävään sopivammaksi
     tätä voisi vielä optimoida jonkin verran paremmaksi mutta saa nyt kelvata tähännn.)*/
-unsigned int kysyPosInt(unsigned int minVal, unsigned long long maxVal){
+unsigned int kysyPosInt(unsigned int minVal, unsigned int maxVal){
     unsigned long long syote = maxVal+1;
     int maxPituus = 25;
     char ch[maxPituus+1];
